+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## version 1.0.2 (2017.01.24)
+
+1. Complete Roster List.
+2. Complete Roster Upload.
+
 ## version 1.0.1 (2017.01.23)
 
 1. Add Roster model.